requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==1.6.0
beautifulsoup4==4.10.0
boto3==1.37.11
botocore==1.37.11
branca==0.8.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
contourpy==1.3.1
cryptography==36.0.1
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==0.5.0
Django==4.2.16
django-admin-volt==1.0.10
django-ajax-datatable==4.4.2
django-allauth==65.4.1
django-computed-property==0.3.0
django-computedfields==0.1.5
django-crispy-forms==1.12.0
django-debug-toolbar==4.2.0
django-environ==0.8.1
django-filter==2.4.0
django-ses==3.1.0
django-tables2==2.3.4
django-tables2-column-shifter==2.0.3
django-widget-tweaks==1.4.9
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
filelock==3.17.0
folium==0.19.5
fonttools==4.56.0
geoip2==2.9.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
maxminddb==2.6.3
numpy==2.2.3
oauthlib==3.2.0
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pbr==6.1.1
pillow==11.1.0
platformdirs==4.3.6
plotly==5.6.0
postgres==4.0
protobuf==6.30.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.3.0
pyparsing==3.0.7
pypdf==5.3.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.5
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.11.4
setuptools==76.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stevedore==5.4.1
str2bool==1.1
svglib==1.5.1
tenacity==9.0.0
tinycss2==1.4.0
toml==0.10.2
tzdata==2025.1
tzlocal==5.3.1
Unipath==1.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
virtualenv==20.29.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.17
xlwt==1.3.0
xyzservices==2025.1.0
